refactor NodeLatencyTracker; change latencies into LoadingCache
