+++ src/main/java/com/datastax/oss/driver/internal/core/loadbalancing/LatencyAndInflightCountLoadBalancingPolicy.java
@@ -33,7 +33,6 @@ import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.Queue;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicReference;
 import javax.annotation.Nullable;
 import net.jcip.annotations.ThreadSafe;
 import org.slf4j.Logger;
@@ -104,7 +103,7 @@ public class LatencyAndInflightCountLoadBalancingPolicy extends DefaultLoadBalan
         };
     latencies = CacheBuilder.newBuilder().weakKeys().build(cacheLoader);
     this.avoidSlowReplicas =
-            profile.getBoolean(DefaultDriverOption.LOAD_BALANCING_POLICY_SLOW_AVOIDANCE, true);
+        profile.getBoolean(DefaultDriverOption.LOAD_BALANCING_POLICY_SLOW_AVOIDANCE, true);
   }
 
   @NonNull
@@ -232,93 +231,16 @@ public class LatencyAndInflightCountLoadBalancingPolicy extends DefaultLoadBalan
     latencies.getUnchecked(node).add(latencyNanos);
   }
 
-  protected static class TimestampedAverage {
-
-    private final long timestamp;
-    protected final long average;
-    private final long nbMeasure;
-
-    TimestampedAverage(long timestamp, long average, long nbMeasure) {
-      this.timestamp = timestamp;
-      this.average = average;
-      this.nbMeasure = nbMeasure;
-    }
-  }
-
-  protected static class NodeLatencyTracker {
-
-    // just for start up
-    private final long thresholdToAccount;
-    private final AtomicReference<TimestampedAverage> current =
-        new AtomicReference<TimestampedAverage>();
-
-    private final long scale = TimeUnit.MILLISECONDS.toNanos(100);
-
-    NodeLatencyTracker(long thresholdToAccount) {
-      this.thresholdToAccount = thresholdToAccount;
-    }
-
-    public void add(long newLatencyNanos) {
-      TimestampedAverage previous, next;
-      do {
-        previous = current.get();
-        next = computeNextAverage(previous, newLatencyNanos);
-      } while (next != null && !current.compareAndSet(previous, next));
-    }
-
-    private TimestampedAverage computeNextAverage(
-        TimestampedAverage previous, long newLatencyNanos) {
-
-      long currentTimestamp = System.nanoTime();
-
-      long nbMeasure = previous == null ? 1 : previous.nbMeasure + 1;
-      if (nbMeasure < thresholdToAccount)
-        return new TimestampedAverage(currentTimestamp, -1L, nbMeasure);
-
-      if (previous == null || previous.average < 0)
-        return new TimestampedAverage(currentTimestamp, newLatencyNanos, nbMeasure);
-
-      // Note: it's possible for the delay to be 0, in which case newLatencyNanos will basically be
-      // discarded. It's fine: nanoTime is precise enough in practice that even if it happens, it
-      // will be very rare, and discarding a latency every once in a while is not the end of the
-      // world.
-      // We do test for negative value, even though in theory that should not happen, because it
-      // seems
-      // that historically there has been bugs here
-      // (https://blogs.oracle.com/dholmes/entry/inside_the_hotspot_vm_clocks)
-      // so while this is almost surely not a problem anymore, there's no reason to break the
-      // computation
-      // if this even happen.
-      long delay = currentTimestamp - previous.timestamp;
-      if (delay <= 0) return null;
-
-      double scaledDelay = ((double) delay) / scale;
-      // Note: We don't use log1p because we it's quite a bit slower and we don't care about the
-      // precision (and since we
-      // refuse ridiculously big scales, scaledDelay can't be so low that scaledDelay+1 == 1.0 (due
-      // to rounding)).
-      double prevWeight = Math.log(scaledDelay + 1) / scaledDelay;
-      long newAverage =
-          (long) ((1.0 - prevWeight) * newLatencyNanos + prevWeight * previous.average);
-
-      return new TimestampedAverage(currentTimestamp, newAverage, nbMeasure);
-    }
-
-    public TimestampedAverage getCurrentAverage() {
-      return current.get();
-    }
-  }
-
   private void compareAndSwapSlowReplica(Object[] currentNodes, int i, int j) {
     NodeLatencyTracker tracker1 = latencies.getUnchecked((Node) currentNodes[i]);
     NodeLatencyTracker tracker2 = latencies.getUnchecked((Node) currentNodes[j]);
     if (tracker1 != null && tracker2 != null) {
-      TimestampedAverage average1 = tracker1.getCurrentAverage();
-      TimestampedAverage average2 = tracker2.getCurrentAverage();
+      NodeLatencyTracker.TimestampedAverage average1 = tracker1.getCurrentAverage();
+      NodeLatencyTracker.TimestampedAverage average2 = tracker2.getCurrentAverage();
       if (average1 != null
           && average2 != null
-          && average1.average > average2.average
-          && System.nanoTime() - average1.timestamp < RETRY_PERIOD) {
+          && average1.getAverage() > average2.getAverage()
+          && System.nanoTime() - average1.getTimestamp() < RETRY_PERIOD) {
         ArrayUtils.swap(currentNodes, i, j);
       }
     }
