+++ src/main/java/com/datastax/oss/driver/internal/core/loadbalancing/LatencySensitiveLoadBalancingPolicy.java
@@ -27,6 +27,9 @@ import com.datastax.oss.driver.internal.core.loadbalancing.helper.MandatoryLocal
 import com.datastax.oss.driver.internal.core.util.ArrayUtils;
 import com.datastax.oss.driver.internal.core.util.collection.QueryPlan;
 import com.datastax.oss.driver.internal.core.util.collection.SimpleQueryPlan;
+import com.datastax.oss.driver.shaded.guava.common.cache.CacheBuilder;
+import com.datastax.oss.driver.shaded.guava.common.cache.CacheLoader;
+import com.datastax.oss.driver.shaded.guava.common.cache.LoadingCache;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
 import java.util.Map;
@@ -35,10 +38,8 @@ import java.util.Queue;
 import java.util.Set;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.ThreadLocalRandom;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicReference;
 import net.jcip.annotations.ThreadSafe;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -85,21 +86,27 @@ public class LatencySensitiveLoadBalancingPolicy extends DefaultLoadBalancingPol
   private static final Logger LOG = LoggerFactory.getLogger(DefaultLoadBalancingPolicy.class);
 
   private static final long NEWLY_UP_INTERVAL_NANOS = MINUTES.toNanos(1);
-  //  private final long SCALE = TimeUnit.MILLISECONDS.toNanos(100);
   private final long THRESHOLD_TO_ACCOUNT = 100;
   private final long RETRY_PERIOD = TimeUnit.SECONDS.toNanos(10);
 
   protected final Map<Node, Long> upTimes = new ConcurrentHashMap<>();
   private final boolean avoidSlowReplicas;
 
-  protected final ConcurrentMap<Node, NodeLatencyTracker> latencies =
-      new ConcurrentHashMap<Node, NodeLatencyTracker>();
+  protected final LoadingCache<Node, NodeLatencyTracker> latencies;
 
   public LatencySensitiveLoadBalancingPolicy(
       @NonNull DriverContext context, @NonNull String profileName) {
     super(context, profileName);
     this.avoidSlowReplicas =
         profile.getBoolean(DefaultDriverOption.LOAD_BALANCING_POLICY_SLOW_AVOIDANCE, true);
+    CacheLoader<Node, NodeLatencyTracker> cacheLoader =
+        new CacheLoader<Node, NodeLatencyTracker>() {
+          @Override
+          public NodeLatencyTracker load(@NonNull Node node) {
+            return new NodeLatencyTracker(THRESHOLD_TO_ACCOUNT);
+          }
+        };
+    latencies = CacheBuilder.newBuilder().weakKeys().build(cacheLoader);
   }
 
   @NonNull
@@ -171,15 +178,15 @@ public class LatencySensitiveLoadBalancingPolicy extends DefaultLoadBalancingPol
           // Reorder the first two replicas in the shuffled list
           // if the first replica's latency is higher than the second replica's latency and
           // the first replica's latency is not too old
-          NodeLatencyTracker tracker1 = latencies.get((Node) currentNodes[0]);
-          NodeLatencyTracker tracker2 = latencies.get((Node) currentNodes[1]);
+          NodeLatencyTracker tracker1 = latencies.getUnchecked((Node) currentNodes[0]);
+          NodeLatencyTracker tracker2 = latencies.getUnchecked((Node) currentNodes[1]);
           if (tracker1 != null && tracker2 != null) {
-            TimestampedAverage average1 = tracker1.getCurrentAverage();
-            TimestampedAverage average2 = tracker2.getCurrentAverage();
+            NodeLatencyTracker.TimestampedAverage average1 = tracker1.getCurrentAverage();
+            NodeLatencyTracker.TimestampedAverage average2 = tracker2.getCurrentAverage();
             if (average1 != null
                 && average2 != null
-                && average1.average > average2.average
-                && System.nanoTime() - average1.timestamp < RETRY_PERIOD) {
+                && average1.getAverage() > average2.getAverage()
+                && System.nanoTime() - average1.getTimestamp() < RETRY_PERIOD) {
               ArrayUtils.swap(currentNodes, 0, 1);
             }
           }
@@ -207,8 +214,7 @@ public class LatencySensitiveLoadBalancingPolicy extends DefaultLoadBalancingPol
       @NonNull DriverExecutionProfile executionProfile,
       @NonNull Node node,
       @NonNull String logPrefix) {
-    latencies.putIfAbsent(node, new NodeLatencyTracker(THRESHOLD_TO_ACCOUNT));
-    latencies.get(node).add(latencyNanos);
+    latencies.getUnchecked(node).add(latencyNanos);
   }
 
   @Override
@@ -219,9 +225,7 @@ public class LatencySensitiveLoadBalancingPolicy extends DefaultLoadBalancingPol
       @NonNull DriverExecutionProfile executionProfile,
       @NonNull Node node,
       @NonNull String logPrefix) {
-    latencies.putIfAbsent(node, new NodeLatencyTracker(THRESHOLD_TO_ACCOUNT));
-    // even if it's DriverTimeoutException, we still want to put the timeout latency (5s) in it
-    latencies.get(node).add(latencyNanos);
+    latencies.getUnchecked(node).add(latencyNanos);
   }
 
   /** Exposed as a protected method so that it can be accessed by tests */
@@ -232,84 +236,5 @@ public class LatencySensitiveLoadBalancingPolicy extends DefaultLoadBalancingPol
   /** Exposed as a protected method so that it can be accessed by tests */
   protected int diceRoll1d4() {
     return ThreadLocalRandom.current().nextInt(4);
-  }
-
-  protected static class TimestampedAverage {
-
-    private final long timestamp;
-    protected final long average;
-    private final long nbMeasure;
-
-    TimestampedAverage(long timestamp, long average, long nbMeasure) {
-      this.timestamp = timestamp;
-      this.average = average;
-      this.nbMeasure = nbMeasure;
-    }
-  }
-
-  protected static class NodeLatencyTracker {
-
-    private final long thresholdToAccount;
-    //    private final double localScale;
-    private final AtomicReference<TimestampedAverage> current =
-        new AtomicReference<TimestampedAverage>();
-
-    private final long scale = TimeUnit.MILLISECONDS.toNanos(100);
-
-    NodeLatencyTracker(long thresholdToAccount) {
-      //      this.localScale = (double) localScale; // We keep in double since that's how we'll use
-      // it.
-      this.thresholdToAccount = thresholdToAccount;
-    }
-
-    public void add(long newLatencyNanos) {
-      TimestampedAverage previous, next;
-      do {
-        previous = current.get();
-        next = computeNextAverage(previous, newLatencyNanos);
-      } while (next != null && !current.compareAndSet(previous, next));
-    }
-
-    private TimestampedAverage computeNextAverage(
-        TimestampedAverage previous, long newLatencyNanos) {
-
-      long currentTimestamp = System.nanoTime();
-
-      long nbMeasure = previous == null ? 1 : previous.nbMeasure + 1;
-      if (nbMeasure < thresholdToAccount)
-        return new TimestampedAverage(currentTimestamp, -1L, nbMeasure);
-
-      if (previous == null || previous.average < 0)
-        return new TimestampedAverage(currentTimestamp, newLatencyNanos, nbMeasure);
-
-      // Note: it's possible for the delay to be 0, in which case newLatencyNanos will basically be
-      // discarded. It's fine: nanoTime is precise enough in practice that even if it happens, it
-      // will be very rare, and discarding a latency every once in a while is not the end of the
-      // world.
-      // We do test for negative value, even though in theory that should not happen, because it
-      // seems
-      // that historically there has been bugs here
-      // (https://blogs.oracle.com/dholmes/entry/inside_the_hotspot_vm_clocks)
-      // so while this is almost surely not a problem anymore, there's no reason to break the
-      // computation
-      // if this even happen.
-      long delay = currentTimestamp - previous.timestamp;
-      if (delay <= 0) return null;
-
-      double scaledDelay = ((double) delay) / scale;
-      // Note: We don't use log1p because we it's quite a bit slower and we don't care about the
-      // precision (and since we
-      // refuse ridiculously big scales, scaledDelay can't be so low that scaledDelay+1 == 1.0 (due
-      // to rounding)).
-      double prevWeight = Math.log(scaledDelay + 1) / scaledDelay;
-      long newAverage =
-          (long) ((1.0 - prevWeight) * newLatencyNanos + prevWeight * previous.average);
-
-      return new TimestampedAverage(currentTimestamp, newAverage, nbMeasure);
-    }
-
-    public TimestampedAverage getCurrentAverage() {
-      return current.get();
-    }
   }
 }
